# docs/openapi.yaml
openapi: 3.0.3
info:
  title: FlashQ - Kafka-like Streaming API
  description: |
    A Kafka-inspired record queue system with offset-based consumption, consumer groups, 
    and streaming semantics while maintaining single-server simplicity.

    ## Features
    - Topic-based record organization
    - Offset-based consumption for record replay
    - Consumer groups for independent consumption tracking
    - Record keys and headers for routing and metadata
    - FIFO ordering guarantees within topics

    ## API Design Principles
    - RESTful endpoints following Kafka conventions
    - JSON request/response format
    - Offset-based addressing for scalable consumption
    - Consumer group coordination for parallel processing
  version: 1.0.0
  contact:
    name: FlashQ
    url: https://github.com/harshitmahapatra/flashq
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080
    description: Development server
  - url: http://localhost:9090
    description: Alternative development server
paths:
...
  /topics/{topic}/records:
    post:
      summary: Produce records to a topic
      description: |
        Publish one or more records to the specified topic. Records are appended 
        to the topic log and assigned sequential offsets.
      operationId: produceRecords
      tags:
        - Producer
      parameters:
        - $ref: "#/components/parameters/TopicName"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProduceRequest"
            examples:
              single-record:
                summary: Single record
                value:
                  records:
                    - key: "user123"
                      value: "User logged in"
                      headers:
                        source: "auth-service"
                        timestamp: "2024-01-15T10:30:00Z"
              multiple-records:
                summary: Multiple records
                value:
                  records:
                    - key: "user123"
                      value: "User logged in"
                      headers:
                        source: "auth-service"
                    - key: "user456"
                      value: "User logged out"
                      headers:
                        source: "auth-service"
      responses:
        "200":
          description: Records successfully produced
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProduceResponse"
              example:
                offset: 142
                timestamp: "2024-01-15T10:30:00.123Z"
        "400":
          $ref: "#/components/responses/BadRequest"
        "422":
          description: Invalid record format
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: "validation_error"
                message: "Record value cannot be empty"
        "500":
          $ref: "#/components/responses/InternalServerError"

  # Consumer Group Management
  /consumer/{group-id}:
    post:
      summary: Create or join consumer group
      description: |
        Create a new consumer group or join an existing one. Returns the consumer 
        group configuration and current topic subscriptions.
      operationId: createConsumerGroup
      tags:
        - Consumer
      parameters:
        - $ref: "#/components/parameters/ConsumerGroupId"
      responses:
        "201":
          description: Consumer group created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConsumerGroupResponse"
              example:
                group_id: "analytics-processors"
        "200":
          description: Joined existing consumer group
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConsumerGroupResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalServerError"

    delete:
      summary: Leave consumer group
      description: |
        Remove the consumer from the group and trigger rebalancing if necessary.
        All committed offsets are preserved.
      operationId: leaveConsumerGroup
      tags:
        - Consumer
      parameters:
        - $ref: "#/components/parameters/ConsumerGroupId"
      responses:
        "204":
          description: Successfully left consumer group
        "404":
          description: Consumer group not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: "group_not_found"
                message: "Consumer group 'analytics-processors' not found"
        "500":
          $ref: "#/components/responses/InternalServerError"

  # Consumer APIs
  /consumer/{group-id}/topics/{topic}:
    get:
      summary: Fetch records by offset or time
      description: |
        Fetch records from the topic starting at the consumer group's current offset, from a specific
        `from_offset`, or from a specific `from_time` (RFC3339). Polling does not advance the group's
        committed offset; use the offset commit endpoint to advance it.
      operationId: fetchRecords
      tags:
        - Consumer
      parameters:
        - $ref: "#/components/parameters/ConsumerGroupId"
        - $ref: "#/components/parameters/TopicName"
        - name: max_records
          in: query
          required: false
          description: "Maximum number of records to return (default: 100)"
          schema:
            type: integer
            minimum: 1
            maximum: 10000
            default: 100
          example: 50
        - name: from_offset
          in: query
          required: false
          description: Seek to specific offset before fetching (overrides current offset)
          schema:
            type: integer
            minimum: 0
          example: 100
        - name: from_time
          in: query
          required: false
          description: |
            Seek to the first record whose timestamp is greater than or equal to this RFC3339 time.
            Mutually exclusive with `from_offset`.
          schema:
            type: string
            format: date-time
          example: "2025-01-01T00:00:00Z"
        - name: include_headers
          in: query
          required: false
          description: "Whether to include record headers in the response (default: true)"
          schema:
            type: boolean
            default: true
          example: true
      responses:
        "200":
          description: Records fetched successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FetchResponse"
              examples:
                with-records:
                  summary: Successful fetch with records
                  value:
                    records:
                      - offset: 142
                        key: "user123"
                        value: "User logged in"
                        timestamp: "2024-01-15T10:30:00.123Z"
                        headers:
                          source: "auth-service"
                      - offset: 143
                        key: "user456"
                        value: "User logged out"
                        timestamp: "2024-01-15T10:30:01.456Z"
                        headers:
                          source: "auth-service"
                    next_offset: 144
                    high_water_mark: 200
                empty-fetch:
                  summary: No new records available
                  value:
                    records: []
                    next_offset: 144
                    high_water_mark: 144
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          description: Consumer group or topic not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: "topic_not_found"
                message: "Topic 'user-events' not found"
        "500":
          $ref: "#/components/responses/InternalServerError"

  # Offset Management
  /consumer/{group-id}/topics/{topic}/offset:
    post:
      summary: Commit offset
      description: |
        Commit the current offset for the consumer group on the specified topic.
        This marks records up to this offset as processed.
      operationId: commitOffset
      tags:
        - Consumer
      parameters:
        - $ref: "#/components/parameters/ConsumerGroupId"
        - $ref: "#/components/parameters/TopicName"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OffsetCommitRequest"
            example:
              offset: 150
              metadata: "Processed batch #42"
      responses:
        "200":
          description: Offset committed successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OffsetCommitResponse"
              example:
                topic: "user-events"
                committed_offset: 150
                timestamp: "2024-01-15T10:35:00.789Z"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          description: Consumer group or topic not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          $ref: "#/components/responses/InternalServerError"

    get:
      summary: Get committed offset
      description: |
        Retrieve the current committed offset for the consumer group on the specified topic.
      operationId: getCommittedOffset
      tags:
        - Consumer
      parameters:
        - $ref: "#/components/parameters/ConsumerGroupId"
        - $ref: "#/components/parameters/TopicName"
      responses:
        "200":
          description: Current committed offset
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OffsetResponse"
              example:
                topic: "user-events"
                committed_offset: 150
                high_water_mark: 200
                lag: 50
                last_commit_time: "2024-01-15T10:35:00.789Z"
        "404":
          description: Consumer group or topic not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          $ref: "#/components/responses/InternalServerError"
components:
  schemas:
...
    ProduceRequest:
      type: object
      required:
        - records
      properties:
        records:
          type: array
          description: |
            Array of records to produce. Limited to 1000 records per batch to prevent 
            memory exhaustion and ensure reasonable response times while supporting 
            high-throughput scenarios.
          minItems: 1
          maxItems: 1000
          items:
            $ref: "#/components/schemas/Record"
    ProduceResponse:
      type: object
      required:
        - offset
        - timestamp
      properties:
        offset:
          type: integer
          description: Assigned offset for the last record in the batch
          minimum: 0
          example: 142
        timestamp:
          type: string
          format: date-time
          description: Server timestamp when the last record was stored
          example: "2024-01-15T10:30:00.123Z"

    # Consumer Group Schemas
    ConsumerGroupResponse:
      type: object
      required:
        - group_id
      properties:
        group_id:
          type: string
          description: The consumer group identifier
          example: "analytics-processors"

    # Fetch Response
    FetchResponse:
      type: object
      required:
        - records
        - next_offset
        - high_water_mark
      properties:
        records:
          type: array
          description: Fetched records with offsets
          items:
            $ref: "#/components/schemas/RecordWithOffset"
        next_offset:
          type: integer
          description: Next offset to fetch from
          minimum: 0
        high_water_mark:
          type: integer
          description: Highest offset in the topic
          minimum: 0
        lag:
          type: integer
          description: Number of records behind high water mark (optional, can be calculated as high_water_mark - next_offset)
          minimum: 0

    # Offset Management
    OffsetCommitRequest:
      type: object
      required:
        - offset
      properties:
        offset:
          type: integer
          description: Offset to commit
          minimum: 0
          example: 150
        metadata:
          type: string
          description: Optional metadata about the commit
          maxLength: 1024
          example: "Processed batch #42"
    OffsetCommitResponse:
      type: object
      required:
        - topic
        - committed_offset
        - timestamp
      properties:
        topic:
          type: string
          description: Topic name
          example: "user-events"
        committed_offset:
          type: integer
          description: The committed offset
          minimum: 0
          example: 150
        timestamp:
          type: string
          format: date-time
          description: When the offset was committed
          example: "2024-01-15T10:35:00.789Z"
    OffsetResponse:
      type: object
      required:
        - topic
        - committed_offset
        - high_water_mark
        - lag
      properties:
        topic:
          type: string
          description: Topic name
          example: "user-events"
        committed_offset:
          type: integer
          description: Current committed offset
          minimum: 0
          example: 150
        high_water_mark:
          type: integer
          description: Highest available offset
          minimum: 0
          example: 200
        lag:
          type: integer
          description: Number of records behind
          minimum: 0
          example: 50
        last_commit_time:
          type: string
          format: date-time
          description: When offset was last committed
          example: "2024-01-15T10:35:00.789Z"

    # Error Response
    ErrorResponse:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
          description: Error code identifier
          example: "validation_error"
        message:
          type: string
          description: Human-readable error message
          example: "Record value cannot be empty"
        details:
          type: object
          description: Additional error details
          additionalProperties: true
  responses:
    BadRequest:
      description: Bad request - invalid parameters or request body
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: "invalid_parameter"
            message: "Topic name must contain only alphanumeric characters, dots, underscores, and hyphens"
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: "internal_error"
            message: "An unexpected error occurred while processing the request"
  parameters:
    TopicName:
      name: topic
      in: path
      required: true
      description: The topic name
      schema:
        type: string
        pattern: "^[a-zA-Z0-9._][a-zA-Z0-9._-]*$"
        minLength: 1
        maxLength: 255
      example: "user-events"
    ConsumerGroupId:
      name: group-id
      in: path
      required: true
      description: The consumer group identifier
      schema:
        type: string
        pattern: "^[a-zA-Z0-9._][a-zA-Z0-9._-]*$"
        minLength: 1
        maxLength: 255
      example: "analytics-processors"
...